Make DualSignal.average() return u16, not u32
There's no way an average of two 16 bit numbers has to be a 32 bit number.

+++ src/dual_signal.rs
@@ -43,8 +43,8 @@ where
         self.high = (high / DAC_SAMPLE_AVERAGE_COUNT) as _;
     }
 
-    pub fn average(&self) -> u32 {
-        (u32::from(self.low) + u32::from(self.high)) / 2
+    pub fn average(&self) -> u16 {
+        ((u32::from(self.low) + u32::from(self.high)) / 2) as u16
     }
 
     pub fn diff(&self) -> u16 {
